ggsignif-package {ggsignif}	R Documentation

ggsignif: Significance Brackets for 'ggplot2'

Description

Enrich your 'ggplots' with group-wise comparisons. This package provides an easy way to indicate if two groups are significantly different. Commonly this is shown by a bracket on top connecting the groups of interest which itself is annotated with the level of significance (NS, *, **, ***). The package provides a single layer (geom_signif()) that takes the groups for comparison and the test (t.test(), wilcox.text() etc.) as arguments and adds the annotation to the plot.

Author(s)

Maintainer: Constantin Ahlmann-Eltze artjom31415@googlemail.com (ORCID) (@const_ae) [contributor]

Authors:

Indrajeet Patil patilindrajeet.science@gmail.com (ORCID) (@patilindrajeets) [contributor]

See Also

Useful links:

https://const-ae.github.io/ggsignif/

https://github.com/const-ae/ggsignif

[Package ggsignif version 0.6.4 Index]